acty-project/backend/ml/rag/requirements.txt
# FSM RAG Pipeline — Python dependencies
# Install with: pip install -r requirements.txt --break-system-packages

# PDF parsing
pymupdf>=1.24.0          # fitz — fast PDF text extraction
pdfplumber>=0.10.0       # table-aware PDF parsing
pytesseract>=0.3.10      # OCR wrapper for Tesseract
Pillow>=10.0.0           # image handling for OCR

# Embeddings & vector store
sentence-transformers>=2.7.0   # BAAI/bge-large-en-v1.5
chromadb>=0.5.0                # persistent local vector DB
torch>=2.2.0                   # GPU acceleration for embeddings

# HTTP (Ollama API)
httpx>=0.27.0
